api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
databricks-sdk==0.12.0
databricks-sql-connector==3.1.0
pandas==2.1.0
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing-extensions==4.7.1
werkzeug==3.0.1
pyyaml==6.0.1
requests==2.31.0 